Add VAT names
To allow adding the local VAT name or abbreviation in invoices, add them to the list.

This allows to show e.g. "Taxes - Umsatzsteuer (USt) - Austria - 20% - 3,33 EUR" or "USt - Austria - 20% - 3,33 EUR" on the invoice, depending on the wishes of invoice creator.

Not every user will understand the English term "Value added tax" or know that his local "Umsatzsteuer" is meant. In addition to being nicer and easier to understand for end users and their accountants, this makes it clearer to the end user that his local tax was applied.

Source: Handelskammer Hamburg (Chamber of commerce of Hamburg, Germany)
https://www.hk24.de/produktmarken/beratung-service/recht_und_steuern/steuerrecht/umsatzsteuer_mehrwertsteuer/umsatzsteuer_mehrwertsteuer_international/warenhandel/umsatzsteuersaetze-eu-drittstaaten/1167702